I created a class player which contains all necessary attributes for a person to play the game such as name, money bet on the game and number of tries..
Two different fucntions were created as well, one for giving a welcome
message and to decide whether the user wants to become a player, and another
(not yet finished) that allows the player to play.

We need to complete the programm by providing controll over the attempts
of the player in order to stop the game when the payer looses and also
over the amount of times the player wants to play the game.