We have included a function that receives two variables, one represents the warrior and the other one represents a dragon. This function represents the battle between the warrior and the dragon for the hidden treasure that the dragon holds.
In here, we set two parameter for Dragon's life and Warrior's life. We
recommend to set Dragon's life higher than Warrior's life.

Also, in this function we base the entire battle on two random variables
for both the warrior and the dragon. If these random variables are not
equal, then the warrior cannot defend himself and therefore, the
warrior receives an attack. On the other hand, if
these random variables are equal, then the warrior defends himself and
attacks.