The following updates were included in the developing of this game:
We included a method for the player that allows us to create the underscore blacks depending on the length of the word hidden, this is done automatically.

A welcoming message was also included as a function. This was done in order to  provide some sense to anyone who might use this program and it will allow that person to see the rules of this game.

Lastly, a function called "CheckingResult" was included and it will allow us to determine if a letter provided by a player is or is not part of the hidden word.