We have created a function that represents the game itself but we have included a few external functionalities that will be used by this function in order to make the game work. These functions are:
CheckingResult: This function will receive two classes and we will check
if a selected letter by the player is part of the selected word by the
hanger in iterative ways until the player guesses the word or looses his
chances.

The other two functions are called "LookInWord" and "ShowResult". The
first one will look if the selected letter by the player is in the word
selected by the hanger and it will fill the underscores that appear in
the game in order to hide the entire word from the player. The other
function just shows the hidden word with the underscores and guessed
letters as well as the amount of chanches that the player still has.